host.c host.h cl_main.c: Host_Frame now takes a double instead of a float. qw/Makefile.am: 	Reformatted the file list. 	Removed gl_refrag.c. Everything else: 	The great entity rewrite! 	I killed efrags, completely.

+++ src/include/host.h
@@ -46,7 +46,7 @@ void Host_Init (void);
 void Host_Shutdown (void);
 void Host_Error (char *error, ...);
 void Host_EndGame (char *message, ...);
-void Host_Frame (float time);
+void Host_Frame (double time);
 void Host_Quit_f (void);
 void Host_ClientCommands (char *fmt, ...);
 void Host_ShutdownServer (qboolean crash);
